% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{filepreds}
\alias{does_file_exist}
\alias{filepreds}
\alias{has_extension}
\alias{is_directory}
\alias{is_readable}
\alias{is_writeable}
\title{File system predicates}
\usage{
does_file_exist(x)

is_directory(x)

is_writeable(x)

is_readable(x)

has_extension(x, ext)
}
\arguments{
\item{x}{file system object to test}

\item{ext}{file extension as a string}
}
\value{
logical value
}
\description{
File system predicates
}
\examples{
does_file_exist(system.file("DESCRIPTION")
is_directory(system.file())
has_extension(system.file("DESCRIPTION"), "")
has_extension(system.file("R/base.rdb"), "rdb")
}

