% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{is_date}
\alias{is_date}
\title{Is a Date?}
\usage{
is_date(x)
}
\arguments{
\item{x}{object to test}
}
\value{
\code{TRUE} if \code{x} is a \code{Date}. Otherwise \code{FALSE}
}
\description{
Is a Date?
}
\examples{
is_date(Sys.Date())
is_date(Sys.time())
}

