% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{is_formula}
\alias{is_formula}
\title{Is a formula?}
\usage{
is_formula(x)
}
\arguments{
\item{x}{object to test}
}
\value{
\code{TRUE} if \code{x} inherits from \code{formula}. Otherwise
  \code{FALSE}
}
\description{
Is a formula?
}
\examples{
x <- disp ~ am
is_formula(x)
}

