% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{is_scalar}
\alias{is_scalar}
\title{Is of length one?}
\usage{
is_scalar(x)
}
\arguments{
\item{x}{object to test}
}
\value{
\code{TRUE} if length of \code{x} is one. Otherwise \code{FALSE}
}
\description{
Is of length one?
}
\examples{
is_scalar(c(1, 2))
is_scalar(1)
}

