% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{is_time}
\alias{is_time}
\title{Is a time (Date-Time)?}
\usage{
is_time(x)
}
\arguments{
\item{x}{object to test}
}
\value{
\code{TRUE} if \code{x} inherits from \code{\link{POSIXt}}.
  Otherwise \code{FALSE}
}
\description{
Is a time (Date-Time)?
}
\examples{
is_time(Sys.Date())
is_time(Sys.time())
}

