% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{nas}
\alias{any_na}
\alias{is_na}
\alias{nas}
\title{NA predicates}
\usage{
is_na(x)

any_na(x)
}
\arguments{
\item{x}{object to test}
}
\description{
\code{is_na} returns \code{TRUE} if \code{x} is an \code{NA} of length one
and otherwise \code{FALSE} (including when \code{x} a vector of more than
one \code{NA}). Meanwhile \code{any_na} returns \code{TRUE} if \code{x}
contains any NAs (but not recursively to lists and pairlists), otherwise
\code{FALSE}.
}
\examples{
is_na(c(1, NA))
is_na(NA)
any_na(c(1, NA))
any_na(list(1, list(1, NA)))
}

