% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{nas}
\alias{any_na}
\alias{is_na}
\alias{nas}
\title{Is \code{NA}?}
\usage{
is_na(x)

any_na(x)
}
\arguments{
\item{x}{object to test}
}
\description{
\code{is_na} returns a logical vector containing \code{TRUE} if \code{x} is
\code{NA} and otherwise \code{FALSE}, while \code{any_na} returns \code{TRUE}
if \code{x} contains any NAs (recursively to lists and pairlists), otherwise
\code{FALSE}.
}
\examples{
is_na(c(1, NA))
any_na(c(1, NA))
any_na(list(1, list(1, NA)))
}

