% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{nas}
\alias{any_na}
\alias{is_na}
\alias{is_scalar_na}
\alias{nas}
\title{NA predicates}
\usage{
is_na(x)

is_scalar_na(x)

any_na(x)
}
\arguments{
\item{x}{object to test}
}
\description{
\code{is_na} returns \code{TRUE} if \code{x} are all \code{NA} and otherwise
\code{FALSE}. Meanwhile \code{any_na} returns \code{TRUE} if \code{x}
contains any NAs (but not recursively to lists and pairlists), otherwise
\code{FALSE}, while \code{is_scalar_na} returns \code{TRUE} if \code{x} is
singular \code{NA}.
}
\examples{
is_na(c(1, NA))
is_na(NA)
any_na(c(1, NA))
any_na(list(1, list(1, NA)))
}

